refactor(http): keep tracer state in middleware
